+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/Control.java
@@ -3132,17 +3132,18 @@ public abstract class Control extends Widget implements Drawable {
 		if (gc.isDisposed())
 			error(SWT.ERROR_INVALID_ARGUMENT);
 
+		NativeGC ngc = (NativeGC) gc.innerGC;
 		if (OS.VERSION < OS.VERSION(10, 13, 0)) {
 			NSGraphicsContext.static_saveGraphicsState();
-			NSGraphicsContext.setCurrentContext(gc.handle);
+			NSGraphicsContext.setCurrentContext(ngc.handle);
 			NSAffineTransform transform = NSAffineTransform.transform();
 			transform.translateXBy(0, view.bounds().height);
 			transform.scaleXBy(1, -1);
 			transform.concat();
-			view.displayRectIgnoringOpacity(view.bounds(), gc.handle);
+			view.displayRectIgnoringOpacity(view.bounds(), ngc.handle);
 			NSGraphicsContext.static_restoreGraphicsState();
 		} else {
-			view.displayRectIgnoringOpacity(view.bounds(), gc.handle);
+			view.displayRectIgnoringOpacity(view.bounds(), ngc.handle);
 		}
 		return true;
 	}
